requirements.txt
pandas==2.1.3
python-dotenv==1.0.1
groq==0.18.0
langchain-huggingface==0.0.6
langchain-community==0.3.18
streamlit==1.42.2
python-docx==1.1.2
openpyxl==3.1.5
